// backend/src/pessoas/interfaces/fisica.out.ts
export interface FisicaOut {
  pes_id: number;
  pes_nome: string;
  pes_email: string;
  pes_telefone: string;
  pes_logradouro: string;
  pes_complemento: string;
  pes_numero: string;
  pes_bairro: string;
  est_sigla: string;
  cid_nome: string;
  pes_status: boolean;
  fis_cpf: string;
  fis_rg: string;
  sex_id: number;
  fis_data_nascimento: Date;
}